MOved RPS Manager to its own module
